move rustyline and ctrlc features in cli.rs to feature.rs
